requirements.txt
Django==1.6.11
djangorestframework==1.0
wheel==0.24.0
